+++ server.js
@@ -1,8 +1,9 @@
 const express = require('express')
 const hbs = require('express-handlebars')
 const server = express()
-// const messageData = require('./message.json')
-const aes_256 = require('./aes-256')
+
+const encryption = require('./routes/encryption')
+const decryption = require('./routes/decryption')
 
 
 server.engine('hbs', hbs({
@@ -10,81 +11,33 @@ server.engine('hbs', hbs({
     extname: 'hbs'
   }))
 
+  server.engine('hbs', hbs({
+    defaultLayout: 'main',
+    extname: 'hbs'
+  }))
+
 server.set('view engine', 'hbs') // allow access to hbs file directory
 server.use(express.static('public')) // allow access to public folder
 server.use(express.urlencoded({extended: false})) //need clarification about this
+server.use('/encryption', encryption)
+server.use('/decryption', decryption)
 
 server.get('/', (req, res) => {
     const home = './home.hbs'
     res.render(home)
 })
 
-server.get('/encryption-demo', (req, res) => {
-    const encryption = './encryption'
-    data = {
-        iv: '1234567812345678',
-        key: '12345678123456781234567812345678',
-        plainTextMessage: "",
-        encryptedMessage: ""
-    }
-    res.render(encryption, data)
+server.get('/encryption', (req, res) => {
+    res.send(encryption)
 })
 
-server.get('/decryption-demo', (req, res) => {
-  const decryption = './decryption'
-  data = {
-    iv: '1234567812345678',
-    key: '12345678123456781234567812345678',
-    plainTextMessage: '',
-    encryptedMessage: ''
-  }
-  res.render(decryption, data)
+server.get('/decryption', (req, res) => {
+  res.send(decryption)
 })
-
-  
-
-server.post('/encryption-demo', (req, res) => {
-    const encryption = './encryption'
-    let plainTextMessage = req.body.plainText
-    let encryptedMessage = aes_256.encrypt(plainTextMessage)
-   
-    const updatedData = {
-        iv: '1234567812345678',
-        key: '12345678123456781234567812345678',
-        plainTextMessage: plainTextMessage,
-        encryptedMessage: encryptedMessage.encryptedData        
-    }
-    res.render(encryption, updatedData)
-})
-
-server.post('/decryption-demo', (req, res) => {
-  const decryption = './decryption'
-  let encryptedMessage = req.body.encryptedText
-  let plainTextMessage = aes_256.decrypt(encryptedMessage)
-
-  const updatedData = {
-    iv: '1234567812345678',
-    key: '12345678123456781234567812345678',
-    plainTextMessage: plainTextMessage,
-    encryptedMessage: encryptedMessage
-  }
-  res.render(decryption, updatedData)
-})
-
 
 server.get('/key-exchange', (req, res) => {
     const keyExchange = './keyExchange'
     res.render(keyExchange)
 })
-
-
-
-
-
-
-
-
-
-
 
 module.exports = server;
