+++ views/partials/navbar.hbs
@@ -5,8 +5,8 @@
   </button>
   <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
     <div class="navbar-nav">
-      <a class="nav-item nav-link" href="/encryption-demo">Encryption Demo</a>
-      <a class="nav-item nav-link" href="/decryption-demo">Decryption Demo</a>
+      <a class="nav-item nav-link" href="/encryption">Encryption Demo</a>
+      <a class="nav-item nav-link" href="/decryption">Decryption Demo</a>
       <a class="nav-item nav-link" href="/key-exchange">Diffie-Hellman Key Exchange</a>
     </div>
   </div>
